Sonar (#830) + CodeQL #826: split cms/experiments/services into package
The single 777-line ``cms/experiments/services.py`` exceeded the
500-line cap (S104). Split it into a package following the same pattern
already used by ``cms/services/`` and ``engine/services/``:

- ``_common``: shared ``_validate_user`` / ``_check_result_type``.
- ``_scripts``: script asset upload / completion / delete flow,
  including ``_inspect_uploaded_script_body``.
- ``_experiments``: experiment list / get / create / start / cancel,
  plus the ``_resolve_experiment_scenario`` /
  ``_validate_script_assignments`` / ``_resolve_experiment_agent``
  helpers with full parameter and return-type annotations against
  ``ScenarioTemplate``, ``ScriptAssignmentInput`` and ``AgentConfig``
  (S6540, S6538).
- ``_artifacts``: artifact and bundle download URL generators.
- ``_scenarios``: scenario-instance helper, including the CodeQL
  py/log-injection (#826) fix: the ``except`` block at the bottom of
  ``get_scenario_instances`` now CR/LF/tab-sanitises ``scenario_id``
  inline (CodeQL recognises the ``.replace`` chain as a sanitiser
  even though it doesn't pattern-match the equivalent
  ``safe_log_value`` helper).
- ``__init__``: re-exports the public service functions plus the
  patchable names tests target via
  ``patch("cms.experiments.services.<name>")`` (``transaction``,
  ``audit_log``, models, S3 helpers, scenario registry helpers,
  ``publish_experiment_event``). Submodules look these up through the
  package at call time so the patches keep working across the split.

While restructuring, fold in the remaining new-code Sonar findings on
the original file: switch ``logger.error`` exception envelopes to
``logger.exception`` (S8572), and let mypy narrow ``payload["s3_key"]``
/ ``payload["file_size"]`` to ``str`` / ``int`` at the boundary.
